requirements.txt
langchain-openai
langchain-community 
langchain
langchain-huggingface
langchain-chroma
gradio
flask
python-dotenv
nltk
chromadb
faiss-cpu
pdfminer.six
PyMuPDF
pypdf
pdfplumber
torchvision
timm
pandas
markdownify
tabulate